:recycle: Refactor rendering and computing of stats code
